Move the function to core/init
